be/pyproject.toml
[project]
name = "be"
version = "1.0.0"
description = "FastAPI backend for Sudar educational platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-dotenv==1.0.0",
    "psycopg2-binary==2.9.9",
    "bcrypt==4.1.1",
    "sqlalchemy==2.0.23",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "email-validator==2.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]
